requirements.txt
behave==1.2.5
PyHamcrest==1.9.0
selenium==3.6.0
fake-factory==0.5.2